app/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
fastapi-restful==0.6.0
loguru==0.7.2
typing-inspect==0.9.0
pydantic==2.8.2
pydantic-settings==2.4.0
motor==3.6.0
bcrypt==4.2.0
python-jose==3.3.0
python-dateutil==2.9.0.post0

# For test
pytest==8.3.2
pytest-asyncio==0.23.8

# Dependency Injector for DI
dependency-injector==4.42.0